% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maps.R
\name{CountMap}
\alias{CountMap}
\title{Python-like Counter dictionary}
\usage{
cl <- DictMap$new()
cl$add("a")
cl$add("a")
cl$get("a") == 2
cl$head(1)
}
\description{
Counter dictionary to count the occurences of a
character string. The class provides additionaly
a function head(top_n) to get the top n occurences.
Also included is a sort function to sort the internal list.
WARNING: This operation cannot be undone.
}
\section{Super class}{
\code{\link[RolliNLP:DictMap]{RolliNLP::DictMap}} -> \code{CountMap}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-add}{\code{CountMap$add()}}
\item \href{#method-put}{\code{CountMap$put()}}
\item \href{#method-order}{\code{CountMap$order()}}
\item \href{#method-head}{\code{CountMap$head()}}
\item \href{#method-clone}{\code{CountMap$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="RolliNLP" data-topic="DictMap" data-id="get">}\href{../../RolliNLP/html/DictMap.html#method-get}{\code{RolliNLP::DictMap$get()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add"></a>}}
\if{latex}{\out{\hypertarget{method-add}{}}}
\subsection{Method \code{add()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CountMap$add(key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{atomic character string}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-put"></a>}}
\if{latex}{\out{\hypertarget{method-put}{}}}
\subsection{Method \code{put()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CountMap$put(key, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{atomic character string}

\item{\code{value}}{any}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-order"></a>}}
\if{latex}{\out{\hypertarget{method-order}{}}}
\subsection{Method \code{order()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CountMap$order()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-head"></a>}}
\if{latex}{\out{\hypertarget{method-head}{}}}
\subsection{Method \code{head()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CountMap$head(top_n)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CountMap$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
