% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NaiveTokenizer.R
\name{getNGrams}
\alias{getNGrams}
\title{N-gram Tokenizer}
\usage{
getNGrams(token_vec, n = 2)
}
\arguments{
\item{token_vec}{vector of atomic character strings}

\item{n}{cardinality of the pairs}
}
\description{
Get n-grams from a vector of tokens, supplied as parameter.
}
\examples{
getNGrams(c("Hello", "World", "!"), n = 2)

}
