% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%cat\%}
\alias{\%cat\%}
\title{Concatenate two Strings}
\usage{
x \%cat\% y
}
\arguments{
\item{x}{left hand side character string}

\item{y}{right hand side character string}
}
\description{
This function is a wraper for the 'paste0' function
and does exactly the same. Put the first string to the
left and the second string to the right.
}
