% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maps.R
\name{WordDict}
\alias{WordDict}
\title{Word dictionary}
\description{
A datastructure to construct a dictionary for words as keys
to numbers as values. This class also provides the method word2vec()
to transform an already tokenized character vector into
a word embeding. The default value for a missing value is 0 here.
Specific editing can be done using the put() function.
}
\examples{
wd <- WordDict$new(lower = TRUE)
wd$add("Hello")
wd$add("World")
wd$get("world") == 2
wd <- WordDict$new(lower = FALSE)
wd$add("Hello")
wd$add("World")
wd$get("world") == 0 # Missing word
wd$get("World") == 2
wd$word2vec(c("Hello", "World")) == c(1,2)
wd$word2vec(c("Hello", "world")) == c(1,0)

}
\section{Super class}{
\code{\link[RolliNLP:DictMap]{RolliNLP::DictMap}} -> \code{WordDict}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{WordDict$new()}}
\item \href{#method-get}{\code{WordDict$get()}}
\item \href{#method-add}{\code{WordDict$add()}}
\item \href{#method-word2vec}{\code{WordDict$word2vec()}}
\item \href{#method-clone}{\code{WordDict$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="RolliNLP" data-topic="DictMap" data-id="items">}\href{../../RolliNLP/html/DictMap.html#method-items}{\code{RolliNLP::DictMap$items()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="RolliNLP" data-topic="DictMap" data-id="put">}\href{../../RolliNLP/html/DictMap.html#method-put}{\code{RolliNLP::DictMap$put()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WordDict$new(lower = F)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get"></a>}}
\if{latex}{\out{\hypertarget{method-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WordDict$get(key)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add"></a>}}
\if{latex}{\out{\hypertarget{method-add}{}}}
\subsection{Method \code{add()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WordDict$add(word)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-word2vec"></a>}}
\if{latex}{\out{\hypertarget{method-word2vec}{}}}
\subsection{Method \code{word2vec()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WordDict$word2vec(words)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WordDict$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
