% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maps.R
\name{DictMap}
\alias{DictMap}
\title{Python-like dictionary}
\usage{
cl <- DictMap$new()
cl$put("a", 2)
cl$get("a")
}
\description{
With this dictionary one can attach data to
a key. The limitation here is that the key
must be an atomic character string. Functions get()
and put() are provided to edit and get informations
about the class.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-get}{\code{DictMap$get()}}
\item \href{#method-put}{\code{DictMap$put()}}
\item \href{#method-clone}{\code{DictMap$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get"></a>}}
\if{latex}{\out{\hypertarget{method-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DictMap$get(key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{atomic character string}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-put"></a>}}
\if{latex}{\out{\hypertarget{method-put}{}}}
\subsection{Method \code{put()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DictMap$put(key, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{atomic character string}

\item{\code{value}}{any}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DictMap$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
