% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%fmt\%}
\alias{\%fmt\%}
\title{Format a string}
\usage{
input_str \%fmt\% vec
}
\description{
A string with marks in the form of "{}" is formated at
these points with the values passed on the right side.
On the right side is to be passed a vector with the
same amount of substituents as the string has "{}" marks.
}
\examples{
"Hello {}. This is supposed to be a {}!" \%fmt\% c("World", "Test")
}
